added i2c semaphore, moved imu functions
updating implementation for final merge
