+++ SensorController/Core/Inc/Sensors/imu.h
@@ -14,12 +14,28 @@
 /* Headers */
 #include "Peripherals/i2c.h"
 /* Macros */
-
+#define LSM6DS3_ADDR 0b1101011 //7th bit is set to SA0 Pin (GND), so we could use two addresses and dynamically change which one
 /* Data Structures */
 
 /* Global Variables */
 
 /* Public Prototypes */
+/* ----- Data processing functions ----- */
+
+/**
+ * @brief Reads acceleration registers and stores register data in struct
+ * @param dev pointer to sensor struct
+ * @param accelData float buffer to hold acceleration data output
+ */
+HAL_StatusTypeDef LSM6DS3_ReadAccel(DAT_SENSOR *dev, float accelData[3]);
+/**
+ * @brief Reads the temperature registers and stores in struct
+ * @param dev pointer to sensor struct
+ */
+HAL_StatusTypeDef LSM6DS3_ReadTemp(DAT_SENSOR * dev);
+
+HAL_StatusTypeDef LSM6DS3_Init(DAT_SENSOR * dev, I2C_HandleTypeDef * i2cHandle);
+
 /**
  * @brief IMU Sensor Test Task
  *
